add author search
