+++ drevo/forms/author_search_form.py
@@ -1,0 +1,29 @@
+from django import forms
+from ..models import AuthorType
+
+from .custom_choice_field import CustomChoiceField
+from .select_with_input import SelectWithInput
+
+
+class AuthorSearchForm(forms.Form):
+    """
+    Форма для фильтрации авторов по критериям.
+    """
+    author_type_choices = [(author_type, author_type)
+                           for author_type in AuthorType.objects.order_by('name').values_list('name', flat=True)]
+
+    # Поиск по имени автора и по сведениям об авторе
+    main_search = forms.CharField(label="",
+                                  max_length=255,
+                                  widget=forms.TextInput(
+                                      attrs={'class': 'form-control',
+                                             'placeholder': 'Основной поиск'}),
+                                  required=False)
+
+    # Тип автора
+    author_type = CustomChoiceField(label="Тип автора",
+                                    choices=author_type_choices,
+                                    widget=SelectWithInput(
+                                        attrs={'class': 'form-control',
+                                               'placeholder': 'Выберите вид знания'}),
+                                    required=False)
