+++ drevo/forms/select_with_input.py
@@ -1,0 +1,6 @@
+from django import forms
+
+
+class SelectWithInput(forms.Select):
+    template_name = 'drevo/forms/select.html'
+    option_template_name = 'drevo/forms/select_option.html'
