+++ drevo/models/ip.py
@@ -1,0 +1,13 @@
+from django.db import models
+
+
+class IP(models.Model):
+    """
+    IP пользавателей
+    """
+    title = 'IP'
+    visits = models.ManyToManyField('Znanie',
+                                    verbose_name='ID',
+                                    blank=True
+                                    )
+    ip = models.CharField(max_length=100)
