+++ drevo/models/visits.py
@@ -1,0 +1,14 @@
+from django.db import models
+from users.models import User
+
+
+class Visits(models.Model):
+    """
+    Просмотры
+    """
+    znanie = models.ForeignKey('Znanie',
+                               models.CASCADE
+                               )
+    user = models.ForeignKey(User,
+                             models.CASCADE
+                             )
