+++ drevo/templates/drevo/comments_form.html
@@ -3,17 +3,14 @@
         <form id="commentForm{{ comment.id }}">
             <input type="hidden" name="parent" id="id_parent" value="{{ comment.id }}">
             <textarea name="content" cols="1" rows="1" maxlength="{{ comment_max_length }}"
-                      placeholder="{% if comment %} Ответить {% else %} Комментировать {% endif %}"
-                      class="form-control mb-2"
-                      style="overflow: hidden; resize: none;" required=""
-                      id="commentTextareaForm{{ comment.id }}" oninput="onInputHandler(this)"
-                      data-basic-height></textarea>
+                placeholder="{% if comment %} Ответить {% else %} Комментировать {% endif %}" class="form-control mb-2"
+                style="overflow: hidden; resize: none;" required="" id="commentTextareaForm{{ comment.id }}"
+                oninput="onInputHandler(this)" data-basic-height></textarea>
             <div class="form-footer d-flex justify-content-end">
                 <span>
                     <span id="commentCharCounter{{ comment.id }}">0</span>&nbsp;/&nbsp;{{ comment_max_length }}
-                    <input type="button" class="send-comment-button btn btn-secondary ms-2"
-                           value="Отправить" id="sendFormButton{{ comment.id }}"
-                           onclick="onButtonSendClick({{ comment.id }})" disabled>
+                    <input type="button" class="send-comment-button btn btn-secondary ms-2" value="Отправить"
+                        id="sendFormButton{{ comment.id }}" onclick="onButtonSendClick({{ comment.id }})" disabled>
                 </span>
             </div>
         </form>
