+++ drevo/templates/drevo/search.html
@@ -7,7 +7,25 @@
 {% block content %}
 
 <div class="container">
-    <h2 class="text-center">{{ title }}</h2>
+    <h6 class="text-center">
+        {% if request.resolver_match.url_name == 'search_knowledge' %}
+        <span class="fs-3">
+            Знания
+        </span>
+        /
+        <a href="{% url 'search_author' %}" class="link-secondary fs-4" style="text-decoration: none;">
+            Авторы
+        </a>
+        {% elif request.resolver_match.url_name == 'search_author' %}
+        <a href="{% url 'search_knowledge' %}" class="link-secondary fs-4" style="text-decoration: none;">
+            Знания
+        </a>
+        /
+        <span class="fs-3">
+            Авторы
+        </span>
+        {% endif %}
+    </h6>
     <div class="row mt-3">
         <dib class="col">
             <form method="get" class="mx-auto d-flex flex-column" style="max-width:500px">
@@ -65,10 +83,16 @@
     {% if page_obj %}
     <div class="row">
         <div class="col">
+            {% if request.resolver_match.url_name == 'search_knowledge' %}
             {% for knowledge in page_obj %}
-            {% include 'drevo/knowledge.html' %}
-            <hr>
+            {% include 'drevo/search_card_knowledge.html' %}
             {% endfor %}
+
+            {% elif request.resolver_match.url_name == 'search_author' %}
+            {% for author in page_obj %}
+            {% include 'drevo/search_card_author.html' %}
+            {% endfor %}
+            {% endif %}
 
         </div>
     </div>
@@ -81,46 +105,6 @@
 
 </div>
 
-<script>
-
-    // Обрабока полей расширенного поиска
-    extendedSearch = {
-        extendedSearchButton: document.getElementById('collapseExtendedSearchButton'),
-        collapseImg: document.getElementById('extendedSearchButtonImg'),
-        extendedSearchArea: document.getElementById('collapseExtendedSearch'),
-        init() {
-            this.extendedSearchButton.addEventListener('click', (event) => this.saveExtendedFlag(event));
-        },
-        // Перехватываем раскрытие расширенного поиска и сохраняем этот факт
-        saveExtendedFlag(event) {
-            let searchField_iWantCollapse = Boolean(localStorage.getItem('isSearchExtanded'));
-            if (searchField_iWantCollapse) {
-                localStorage.removeItem('isSearchExtanded');
-                this.resetExtendedSearchFields();
-            } else {
-                localStorage.setItem('isSearchExtanded', 'true');
-            }
-        },
-        // Восстанавливаем состояние полей поиска после перезагрузки страницы
-        restoreConditionSearchArea() {
-            if (localStorage.getItem('isSearchExtanded') === 'true') {
-                this.extendedSearchArea.classList.add('show');
-            }
-        },
-        // Сбрасываем значения полей если схлопываем расширенный поиск
-        resetExtendedSearchFields() {
-            let fields = this.extendedSearchArea.querySelectorAll('.form-control');
-            for (let field of fields) {
-                field.value = '';
-            }
-        },
-    }
-
-    extendedSearch.init();
-    extendedSearch.restoreConditionSearchArea();
-
-
-
-
+<script src="{% static 'drevo/js/extended_search.js' %}">
 </script>
 {% endblock content %}
