+++ drevo/templates/drevo/search_card_knowledge.html
@@ -1,21 +1,16 @@
 {% load static %}
 <div class="row">
-    <a class="text-decoration-none" href="{% url 'zdetail' pk=knowledge.pk %}">
-        <h5 class="col-8">
-            {{ knowledge.name }}
-        </h5>
-    </a>
-</div>
-<div class="row">
-
-    <div class="col" style="word-wrap: break-word; width: 100px">
-        {{ knowledge.content | safe | truncatewords:30}}
+    <div class="col-12 col-lg-9">
+        <a class="text-decoration-none" href="{% url 'zdetail' pk=knowledge.pk %}">
+            <h5>
+                {{ knowledge.name }}
+            </h5>
+        </a>
+        <div>
+            {{ knowledge.content | safe | truncatewords_html:30}}
+        </div>
     </div>
-
-</div>
-
-<div class="row">
-    <div class="col-12">
+    <div class="col-12 col-lg-3">
         <ul class="text-end" style="list-style-type: none;">
             <li>
                 Вид знания:
@@ -25,14 +20,12 @@
                 –
                 {% endif %}
             </li>
+            {% if knowledge.category.name %}
             <li>
                 Категория знания:
-                {% if knowledge.category.name %}
                 {{ knowledge.category.name }}
-                {% else %}
-                –
-                {% endif %}
             </li>
+            {% endif %}
             <li>
                 Автор:
                 {% if knowledge.author.name %}
@@ -41,19 +34,19 @@
                 –
                 {% endif %}
             </li>
+            {% if knowledge.related.all %}
             <li>
                 Вид связи:
-                {% if knowledge.related.all %}
                 {% for relations in knowledge.related.all %}
                 {{ relations.tr.name }}
                 {% endfor %}
-                {% else %}
-                –
-                {% endif %}
             </li>
+            {% endif %}
             <li class="d-flex flex-wrap justify-content-end">
                 Изменено: {{ knowledge.updated_at |timesince }} назад
             </li>
         </ul>
     </div>
 </div>
+<hr class="mt-0">
+<sc>
