+++ drevo/views/__init__.py
@@ -10,19 +10,21 @@ from .znanie_by_label_view import ZnanieByLabelView
 from .znanie_detail_view import ZnanieDetailView
 from .znanie_rating_view import ZnanieRatingView
 from .knowledge_search_view import KnowledgeSearchView
+from .author_search_view import AuthorSearchView
 
 
 __all__ = [
-    'author_detail_view',
-    'authors_list_view',
-    'comment_page_view',
-    'comment_send_view',
-    'drevo_list_view',
-    'drevo_view',
-    'glossary_list_view',
-    'labels_list_view',
-    'znanie_by_label_view',
-    'znanie_detail_view',
-    'znanie_rating_view',
-    'knowledge_search_view'
+    'AuthorDetailView',
+    'AuthorsListView',
+    'CommentPageView',
+    'CommentSendView',
+    'DrevoListView',
+    'DrevoView',
+    'GlossaryListView',
+    'LabelsListView',
+    'ZnanieByLabelView',
+    'ZnanieDetailView',
+    'ZnanieRatingView',
+    'KnowledgeSearchView',
+    'AuthorSearchView'
 ]
