+++ drevo/views/labels_list_view.py
@@ -16,6 +16,5 @@ class LabelsListView(ListView):
     template_name = 'drevo/labels.html'
     model = Label
     context_object_name = 'labels'
-    queryset = Label.objects.annotate(zn_num=Count('znanie',
-                                                   filter=Q(znanie__is_published=True))).\
-        all().order_by('name')
+    queryset = Label.objects.annotate(zn_num=Count(
+        'znanie', filter=Q(znanie__is_published=True))).filter(zn_num__gte=1).order_by('name')
